Fix es.po issue causing broken LyXHTML export
Because of a missing backslash, a layout label was not correct.
This caused display issues for the layout in the LyX GUI, but since
the label is not used in export to LaTeX, the ctests for LaTeX
export all passed. Luckily, our LyXHTML test (in particular, the
xmllint part) caught the failure because the lables are exported.

The GUI label now looks good and the following test passes:

    export/examples/es/linguistics_xhtml