Hebrew tutorial: update link to Hebrew set up
Dekel Tsur has a webpage with information on getting set up with
using Hebrew with LaTeX and LyX. We link to that webpage from our
Hebrew Tutorial. He appears to have changed universities, and the
new link gives the same information as the old link did (verified
with web.archive.org).