Collapsable -> Collapsible (part 3)
The current spelling is not strictly wrong, but flagged as unusual or
historical by some authorities. It is also found fault with many
spell checkers. Thus we decided to move to the more standard "-ible"
form once and for all.

See #10678 for discussion

This part addresses the po files.

This will all also all be backported to 2.3.x, for the sake of backwards
compatibility (cherry-picking).