Double escape binary regex in lyx2lyx and fix line-ending problem.
When we switched from r'...' to b'...', this required strings used
as input to re.compile to be double-escaped. It's not an issue with
"\s", because "\s" has no special meaning in a string.

Also, printing in binary mode did not output a line ending.

(cherry picked from commit 6b5a7116cd6697d1121a5ffb76d54102a8e265ba)