Use re.match to find regexp at begin of string/line.
Consistently use the re.match() or the match() method of a
regexp object to find a match at the begin of a line.
This is faster than re.find('^...').

Cf. https://stackoverflow.com/questions/180986/what-is-the-difference-between-re-search-and-re-match
(The subtle difference for multi-line strings is irrelevant for
configure.py, as we match line-wise.)