Try to fix bug #10989.
The problem is that popping dialogs during reload can cause paint
events for which we are not ready. If this does not work, then we
can introduce a new flag, besides 'busy', for that case. But busy
does not seem to be used very widely, so hopefully this works.

Note that when paintEvent is triggered on macOS, the least we can do
is to copy our backing store to the screen.

(cherry picked from commit 19e6977b5b527ec8311da35d8f0a40d6fd509080)
(cherry picked from commit d03deeca0c666256ea68087c5c70b8de73b66edc)