#11142 correct list of previous versions to check for user directory contents
LyX on Mac uses a user directory with version suffix. On change of the version suffix
the existence of the directories with previous versions is checked and the latest one
is used for a copy on first configure run.

For 2.3 the candidate list starts with 2.2 now as it should.

(cherry picked from commit 17c3617c49487977e5c46de20cb450952c68b03d)