Fix bug #6096.
We used to need a hack to set the size of the layout combo, but
the code was changed in Qt 4.5 or so. Hence the appearance of this
bug in 2009. We can now just remove this hack, and all seems to
work correctly.

(cherry picked from commit d7f4346a19b69ae4f46ceba40b9e3a9744bb33f6)